refactor(route): decouple the route operator from controlplane/ffi
- Move the hardware route identity into a stdlib-only leaf package under the
  route control plane and keep a type alias in place, so the module's public
  API and its internal uses are unchanged.
- Point the route operator's neighbour entry at the leaf package: the operator
  no longer links the cgo shared-memory stack and builds with CGO_ENABLED=0.
- Drop the hand-maintained cgo linker depends from the operator's meson target.

Closes #2083.
