+++ modules/route/controlplane/backend.go
@@ -1,14 +1,12 @@
 package route
 
 import (
-	"bytes"
-	"cmp"
 	"fmt"
-	"net"
 
 	"github.com/yanet-platform/yanet2/common/go/bitset"
 	"github.com/yanet-platform/yanet2/controlplane/ffi"
 	"github.com/yanet-platform/yanet2/modules/route/bindings/go/croute"
+	"github.com/yanet-platform/yanet2/modules/route/controlplane/hwroute"
 	"github.com/yanet-platform/yanet2/modules/route/controlplane/routepb/v1"
 )
 
@@ -216,37 +214,12 @@ func (m *backend) counters(
 	return views
 }
 
-// HardwareRoute represents a route in the Layer 2 (L2) networking stack.
+// HardwareRoute is the dataplane's Layer 2 forwarding identity.
 //
-// This is the dataplane's forwarding identity: two nexthops agreeing on
-// these three fields share one route slot regardless of what else the
-// request says about them (e.g. counter name). Keying UpdateModule's dedup
-// map on more would split one physical neighbour across slots, skewing ECMP.
-type HardwareRoute struct {
-	// SourceMAC is the MAC address of the local interface that observed
-	// the neighbour.
-	SourceMAC [6]byte
-	// DestinationMAC is the MAC address of the next hop.
-	DestinationMAC [6]byte
-	// Device is the interface name.
-	Device string
-}
-
-func (m HardwareRoute) String() string {
-	return fmt.Sprintf("%s -> %s", net.HardwareAddr(m.SourceMAC[:]), net.HardwareAddr(m.DestinationMAC[:]))
-}
-
-// Compare compares two hardware routes lexicographically for deterministic sorting.
-func (m HardwareRoute) Compare(other HardwareRoute) int {
-	if c := bytes.Compare(m.SourceMAC[:], other.SourceMAC[:]); c != 0 {
-		return c
-	}
-	if c := bytes.Compare(m.DestinationMAC[:], other.DestinationMAC[:]); c != 0 {
-		return c
-	}
-
-	return cmp.Compare(m.Device, other.Device)
-}
+// The type lives in the leaf hwroute package so that consumers needing only
+// the identity, such as the route operator, do not link the cgo shared-memory
+// stack behind this package; the alias keeps this package's public API intact.
+type HardwareRoute = hwroute.HardwareRoute
 
 func newHardwareRoute(nh *routepb.FIBNexthop) (HardwareRoute, error) {
 	src := nh.GetSrcMac()
